chore: move tx type to separate component
