+++ apps/frontend/src/app/2_molecules/TransactionTypeRenderer/TransactionTypeRenderer.tsx
@@ -1,0 +1,40 @@
+import React, { FC, useCallback } from 'react';
+
+import { useTranslation } from 'react-i18next';
+
+import { translations } from '../../../locales/i18n';
+import { StabilityDepositOperation } from '../../../utils/graphql/zero/generated';
+
+type TransactionTypeRendererProps = {
+  type: StabilityDepositOperation;
+};
+
+export const TransactionTypeRenderer: FC<TransactionTypeRendererProps> = ({
+  type,
+}) => {
+  const { t } = useTranslation();
+  const getTransactionType = useCallback(
+    (operation: StabilityDepositOperation) => {
+      switch (operation) {
+        case StabilityDepositOperation.DepositTokens:
+          return t(
+            translations.stabilityPoolHistory.stabilityPoolOperation.deposit,
+          );
+        case StabilityDepositOperation.WithdrawTokens:
+          return t(
+            translations.stabilityPoolHistory.stabilityPoolOperation.withdrawal,
+          );
+        case StabilityDepositOperation.WithdrawCollateralGain:
+          return t(
+            translations.stabilityPoolHistory.stabilityPoolOperation
+              .withdrawCollateralGain,
+          );
+        default:
+          return '';
+      }
+    },
+    [t],
+  );
+
+  return <>{getTransactionType(type)}</>;
+};
